backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
requests==2.31.0
apscheduler==3.10.4
cryptography==42.0.2
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
statsmodels==0.14.1
pandas==2.2.0
scipy==1.12.0
prophet==1.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
